refactor(auth): move JWT expiration to environment variables

+++ src/main/java/com/stockChecker/live_stock_checker/security/JWT/JwtUtils.java
@@ -26,7 +26,7 @@ import java.util.Date;
 public class JwtUtils {
 
     @Value("${spring.app.jwtExpirationsMS}")
-    private Integer jwtExpirations;
+    private long jwtExpirations;
 
     @Value("${spring.app.jwtSecretKey}")
     private String jwtSecretKey;
